Split inflight watcher state helpers
